refactor(routes): extract snapshot routes
